+++ certs.go
@@ -1,19 +1,22 @@
 package empire
 
-import "golang.org/x/net/context"
+import (
+	"github.com/jinzhu/gorm"
+	"golang.org/x/net/context"
+)
 
 type certsService struct {
 	*Empire
 }
 
-func (s *certsService) CertsAttach(ctx context.Context, app *App, cert string) error {
+func (s *certsService) CertsAttach(ctx context.Context, db *gorm.DB, app *App, cert string) error {
 	app.Cert = cert
 
-	if err := s.store.AppsUpdate(app); err != nil {
+	if err := appsUpdate(db, app); err != nil {
 		return err
 	}
 
-	if err := s.releaser.ReleaseApp(ctx, app); err != nil {
+	if err := s.releases.ReleaseApp(ctx, db, app); err != nil {
 		if err == ErrNoReleases {
 			return nil
 		}