+++ ports.go
@@ -14,50 +14,33 @@ type Port struct {
 
 var ErrNoPorts = errors.New("no ports avaiable")
 
-func (s *store) PortsFindOrCreateByApp(app *App) (*Port, error) {
-	p, err := s.PortsFindByApp(app)
+func portsFindOrCreateByApp(db *gorm.DB, app *App) (*Port, error) {
+	p, err := portsFindByApp(db, app)
 
 	// If an error occurred or we found a port, return.
 	if err != nil || p != nil {
 		return p, err
 	}
 
-	return s.PortsAssign(app)
+	return portsAssign(db, app)
 }
 
-func (s *store) PortsFindByApp(app *App) (*Port, error) {
-	return portsFindByApp(s.db, app)
-}
-
-func (s *store) PortsAssign(app *App) (*Port, error) {
+func portsAssign(db *gorm.DB, app *App) (*Port, error) {
 	var port *Port
 
-	t := s.db.Begin()
-
-	port, err := portsFindAvailable(t)
+	port, err := portsFindAvailable(db)
 	if err != nil {
-		t.Rollback()
 		return port, err
 	}
 
 	// Assign app to port
 	port.AppID = &app.ID
 
-	if err := portsUpdate(t, port); err != nil {
-		t.Rollback()
-		return port, err
-	}
-
-	if err := t.Commit().Error; err != nil {
-		t.Rollback()
+	if err := portsUpdate(db, port); err != nil {
 		return port, err
 	}
 
 	return port, nil
-}
-
-func (s *store) PortsUnassign(app *App) error {
-	return portsUnassign(s.db, app)
 }
 
 func portsFindByApp(db *gorm.DB, app *App) (*Port, error) {